# docs/openapi-generated.yaml
openapi: 3.1.0
info:
  title: 'Bar Assistant API'
...
  contact:
    name: 'Bar Assistant'
    url: 'https://barassistant.app'
    email: info@barassistant.app
  license:
    name: MIT
    url: 'https://github.com/karlomikus/bar-assistant/blob/master/LICENSE'
  version: '{{VERSION}}'
servers:
  -
    url: 'http://localhost:8000/api'
    description: 'Local docker development'
  -
    url: 'https://api.barassistant.app/api'
    description: Production
paths:
...
  /cocktail-methods:
    get:
      tags:
        - 'Cocktail method'
      summary: 'List methods'
      description: 'Show a list of all cocktail methods in a bar'
      operationId: listCocktailMethods
      parameters:
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/CocktailMethod' } }
                type: object
    post:
      tags:
        - 'Cocktail method'
      summary: 'Create method'
      description: 'Create a new cocktail method'
      operationId: saveCocktailMethod
      parameters:
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CocktailMethodRequest'
      responses:
        '201':
          description: 'Successful response'
          headers:
            Location:
              description: 'URL of the new resource'
              schema:
                type: string
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/CocktailMethod' }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
  '/cocktail-methods/{id}':
    get:
      tags:
        - 'Cocktail method'
      summary: 'Show method'
      description: 'Show a specific cocktail method'
      operationId: showCocktailMethod
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/CocktailMethod' }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
    put:
      tags:
        - 'Cocktail method'
      summary: 'Update method'
      description: 'Update a specific cocktail method'
      operationId: updateCocktailMethod
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CocktailMethodRequest'
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/CocktailMethod' }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
    delete:
      tags:
        - 'Cocktail method'
      summary: 'Delete method'
      description: 'Delete a specific cocktail method'
      operationId: deleteCocktailMethod
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Successful response'
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
...
components:
  schemas:
...
    APIError:
      required:
        - type
        - message
      properties:
        type:
          type: string
          example: api_error
        message:
          type: string
          example: 'Resource record not found.'
      type: object
...
    CocktailMethod:
      required:
        - id
        - name
        - dilution_percentage
        - cocktails_count
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Shake
        dilution_percentage:
          type: integer
          example: 20
        cocktails_count:
          type: integer
          example: 32
      type: object
    CocktailMethodRequest:
      required:
        - name
        - dilution_percentage
      properties:
        name:
          type: string
          example: Shake
        dilution_percentage:
          type: integer
          example: 20
      type: object
...
  securitySchemes:
    user_token:
      type: http
      scheme: bearer
security:
  -
    user_token: []
...